pyproject.toml
[project]
name = "lip-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-play-scraper>=1.2.7",
    "groq>=1.2.0",
    "hdbscan>=0.8.42",
    "jinja2>=3.1.6",
    "keybert>=0.9.0",
    "langdetect>=1.0.9",
    "numpy>=2.4.4",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "sentence-transformers>=5.4.1",
    "structlog>=25.5.0",
    "typer>=0.24.2",
    "tzdata>=2026.2",
    "umap-learn>=0.5.12",
]
